+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- TcpServer task not being stopped before the derived object is destroyed.
+
 ## [1.3.3] - 2026-08-19
 ### Fixed
 - EspNetworkInterface IPv4 events not filtered by netif and Ethernet IP-lost event ignored.